Improved performance of producer
The FeedParser now stops the producer the second it finds an article that it has already seen. The assumption is that feeds are ordered by date. The second that I see a story a second time then I can assume that all the following articles will have been seen as well.